Agent_week3_RAG_search/requirements.txt
openai>=1.40.0
chromadb>=0.5.5
pypdf>=4.2.0
tiktoken>=0.7.0
python-dotenv>=1.0.1
tavily-python>=0.3.6

